requirements.txt
boto3>=1.28.0
anthropic>=0.39.0
python-dotenv>=1.0.0
opensearch-py>=2.3.0
